Move the PIM tests in their own module
